# [2.0] Improve URIHelper::parse_url