--- a/src/UriHelper.php
+++ b/src/UriHelper.php
@@ -27,22 +27,54 @@ class UriHelper
 	 * @link    https://secure.php.net/manual/en/function.parse-url.php
 	 * @since   1.0
 	 */
-	public static function parse_url($url)
+	public static function parse_url($url, int $component = -1)
 	{
-		// Get the current LC_CTYPE.
-		$currentLcCType = @setlocale(LC_CTYPE, '0');
+		// Get the current LC_CTYPE locale.
+		$currentLocaleLcCType = setlocale(LC_CTYPE, 0);
 
-		// For utf-8 LC_CTYPE just use PHP native method.
-		if (stripos($currentLcCType, 'utf-8') !== false)
+		if ($currentLocaleLcCType !== false)
 		{
-			return parse_url($url);
+			// UTF-8 LC_CTYPE locale, just use PHP native method.
+			if (stripos($currentLocaleLcCType, 'UTF-8') !== false || stripos($currentLocaleLcCType, 'UTF8') !== false)
+			{
+				return parse_url($url, $component);
+			}
+
+			// Non UTF-8 LC_CTYPE locale, change to C locale before parsing.
+			if (setlocale(LC_CTYPE, 'C') === 'C')
+			{
+				$parsedUrl = parse_url($url, $component);
+				setlocale(LC_CTYPE, $currentLocaleLcCType);
+
+				return $parsedUrl;
+			}
 		}
 
-		// For non utf-8 LC_CTYPE change the LC_CTYPE locale before parsing.
-		@setlocale(LC_CTYPE, 'C');
-		$parsedUrl = parse_url($url);
-		@setlocale(LC_CTYPE, $currentLcCType);
+		// Unable to get LC_CTYPE locale, use old method.
+		$result = false;
 
-		return $parsedUrl;
+		// Build arrays of values we need to decode before parsing
+		$entities     = ['%21', '%2A', '%27', '%28', '%29', '%3B', '%3A', '%40', '%26', '%3D', '%24', '%2C', '%2F', '%3F', '%23', '%5B', '%5D'];
+		$replacements = ['!', '*', "'", "(", ")", ";", ":", "@", "&", "=", "$", ",", "/", "?", "#", "[", "]"];
+
+		// Create encoded URL with special URL characters decoded so it can be parsed
+		// All other characters will be encoded
+		$encodedURL = str_replace($entities, $replacements, urlencode($url));
+
+		// Parse the encoded URL
+		$encodedParts = parse_url($encodedURL, $component);
+
+		// Now, decode each value of the resulting array
+		if ($encodedParts)
+		{
+			$result = [];
+
+			foreach ($encodedParts as $key => $value)
+			{
+				$result[$key] = urldecode(str_replace($replacements, $entities, $value));
+			}
+		}
+
+		return $result;
 	}
 }