--- a/src/UriHelper.php
+++ b/src/UriHelper.php
@@ -41,7 +41,8 @@ class UriHelper
 		$currentLocaleLcCType = @setlocale(LC_CTYPE, 0);
 
 		// If UTF-8 locale found, just use PHP native method which is faster.
-		if ($currentLocaleLcCType !== false && $currentLocaleLcCType === 'C' || stripos($currentLocaleLcCType, 'UTF-8') !== false || stripos($currentLocaleLcCType, 'UTF8') !== false)
+		if ($currentLocaleLcCType !== false && $currentLocaleLcCType === 'C'
+		    || stripos($currentLocaleLcCType, 'UTF-8') !== false || stripos($currentLocaleLcCType, 'UTF8') !== false)
 		{
 			return parse_url($url, $component);
 		}