--- a/src/UriHelper.php
+++ b/src/UriHelper.php
@@ -26,7 +26,6 @@ class UriHelper
 	 * @return  mixed  Associative array or false if badly formed URL.
 	 *
 	 * @link    https://secure.php.net/manual/en/function.parse-url.php
-	 * @link    https://secure.php.net/manual/en/function.setlocale.php
 	 * @since   1.0
 	 */
 	public static function parse_url($url, $component = -1)