Pull Request for Improvement

### Introduction

The `URIHelper::parse_url()` is used in all `JRoute::_()` calls in the cms (ex: building menus, breadcrumbs, sef, etc) so is a very used method.
Also one as to take in consideration that most cms URLs passing trough cms `JRoute::_()` are not utf-8 (ex: `index.php?Itemid=222`)
Improving it's performance will do is little part in improving performance in the CMS.

So this PR tries to improve it's performance by only using the custom parse_url function when the url has utf-8 chars.

### Summary of Changes

Improves performance of URIHelper::parse_url.
Also adds the component parameter as native parse_url.

### Testing Instructions

- Unit tests pass
- Code review
- To test use the following code in a cms 4.0 template index.php before and after the PR
```php
use \Joomla\Uri\UriHelper;

$iterationsPerUrl = 1000;
$uriToTest = [
	'index.php',
	'index.php?ItemId=122',
	'/index.php/username-reminder?view=remind',
	'https://domain.tld/',
	'https://domain.tld/en/about?a=b#anchor',
	'https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor',
	'https://do身ain.tld/中文/纹身馆简介你好/媒体报道',
];


foreach ($uriToTest as $url)
{
	$start = microtime(true);
	for ($i = 1; $i <= $iterationsPerUrl; $i++)
	{
		UriHelper::parse_url($url);
	}
	echo '- [' . number_format(((microtime(true) - $start) * 1000), 3) . ' ms] Processing ' . $url . '' .PHP_EOL;
}
```

My best results on PHP 7.0.19 in Linux for this urls parsed:
- Before:
```
- [2.133 ms] Processing index.php
- [3.315 ms] Processing index.php?ItemId=122
- [3.894 ms] Processing /index.php/username-reminder?view=remind
- [4.136 ms] Processing https://domain.tld/
- [6.849 ms] Processing https://domain.tld/en/about?a=b#anchor
- [10.411 ms] Processing https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor
- [11.326 ms] Processing https://do身ain.tld/中文/纹身馆简介你好/媒体报道
```
- After:
```
- [0.566 ms] Processing index.php
- [0.826 ms] Processing index.php?ItemId=122
- [0.943 ms] Processing /index.php/username-reminder?view=remind
- [0.984 ms] Processing https://domain.tld/
- [1.322 ms] Processing https://domain.tld/en/about?a=b#anchor
- [1.850 ms] Processing https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor
- [4.407 ms] Processing https://do身ain.tld/中文/纹身馆简介你好/媒体报道
```
### Documentation Changes Required

None.

@frankmayer can you also review this.