@andrepereiradasilva 2017-06-25
does anyone as Mac and/or Windows to test if all works in those systems?

In Linux this seems to work fine and tests work fine too.