@andrepereiradasilva 2017-06-27
removed all setlocale stuff, now just check is string has utf8 chars