@frankmayer 2017-06-27
Looks fine on code review.
Sorry, I haven't gotten to setup a Joomla 4 Test environment yet. I can look into that after a few days, if there is still need for that.
BTW, nice improvement. will surely do its part in overall performance of Joomla.
Thanks