@andrepereiradasilva 2017-06-27
made an additional improvement for utf-8 urls (the ones that where still slow)

see https://github.com/joomla-framework/uri/pull/20/commits/8d48df801a4130fe9c95f4151f7a4ab5efe89695 and please review

with this even those utf-8 urls, when mbstring is available (most cases), goes down from 12ms to 4ms