@frankmayer 2017-06-27
strtr in this case is twice as fast as str_replace. Nice one.