@andrepereiradasilva 2017-06-28
work is finished. waiting for reviews
notice in the last commit i removed what i think it's code that as far as i see it was unused. check carefully