@frankmayer 2017-06-28
@andrepereiradasilva just a suggestion. Why don't you include the testing instructions (code) from your initial comment into https://github.com/joomla-framework/uri/blob/master/Tests/UriTest.php? Then it would be better (if not fully) tested?