@andrepereiradasilva 2017-06-28
this is for 2.0 branch and the tests are in another file https://github.com/joomla-framework/uri/blob/2.0-dev/tests/UriHelperTest.php#L15