@mbabker 2017-06-28
If we're going to have benchmarking tests, I think we'd need those in a separate test suite and not something that runs against every commit.  We definitely need benchmarking tests, but they don't have to be as involved in the dev process as unit tests (this can be something we set up in our new CI pipeline, run it once a day/week/month, and make the report available off one of our subdomains).