@frankmayer 2017-06-28
@mbabker I meant only the testing part, not the benchmarking part, but I totally agree.