@mbabker 2017-06-28
On my MacBook...

Before:
```sh
Michaels-MacBook-Pro:uri mbabker$ php tester.php 
- [6.847 ms] Processing index.php
- [9.263 ms] Processing index.php?ItemId=122
- [10.075 ms] Processing /index.php/username-reminder?view=remind
- [13.252 ms] Processing https://domain.tld/
- [18.129 ms] Processing https://domain.tld/en/about?a=b#anchor
- [23.142 ms] Processing https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor
- [24.748 ms] Processing https://do身ain.tld/中文/纹身馆简介你好/媒体报道
```

After:
```sh
Michaels-MacBook-Pro:uri mbabker$ php tester.php 
- [6.320 ms] Processing index.php
- [6.272 ms] Processing index.php?ItemId=122
- [6.073 ms] Processing /index.php/username-reminder?view=remind
- [6.156 ms] Processing https://domain.tld/
- [5.954 ms] Processing https://domain.tld/en/about?a=b#anchor
- [8.268 ms] Processing https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor
- [26.256 ms] Processing https://do身ain.tld/中文/纹身馆简介你好/媒体报道
```