@mbabker 2017-06-28
And on an Ubuntu 16.04 test server:

Before:
```sh
- [2.474 ms] Processing index.php
- [3.704 ms] Processing index.php?ItemId=122
- [4.367 ms] Processing /index.php/username-reminder?view=remind
- [4.660 ms] Processing https://domain.tld/
- [7.375 ms] Processing https://domain.tld/en/about?a=b#anchor
- [11.584 ms] Processing https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor
- [12.306 ms] Processing https://do身ain.tld/中文/纹身馆简介你好/媒体报道
```

After:
```sh
- [1.008 ms] Processing index.php
- [1.087 ms] Processing index.php?ItemId=122
- [1.162 ms] Processing /index.php/username-reminder?view=remind
- [1.317 ms] Processing https://domain.tld/
- [1.787 ms] Processing https://domain.tld/en/about?a=b#anchor
- [2.240 ms] Processing https://domain.tld/path/script.php?foo=%21%2A%27%28%29%3B%3A%40%26%3D%24%2C%2F%3F%23%5B%5D&bar=foo#anchor
- [5.499 ms] Processing https://do身ain.tld/中文/纹身馆简介你好/媒体报道
```